Separate stream_tools + add postgres_tools
