+++ utils/filtered_stream.py
@@ -6,12 +6,14 @@ from dotenv import load_dotenv
 import time
 import yaml
 import csv
+import stream_tools as st
 load_dotenv()
 
 bearer_token = os.environ.get("TWITTER_BEARER_TOKEN")
 update_flag = False
 remove_flag = False
 df = pd.DataFrame()
+export_df = pd.DataFrame()
 
 
 with open("utils/yamls/config.yml", "r") as file:
@@ -20,183 +22,13 @@ with open("utils/yamls/config.yml", "r") as file:
     config["RECONNECT_COUNT"] = 0
 
 
-def bearer_oauth(r):
-    """
-    Method required by bearer token authentication.
-    """
-
-    r.headers["Authorization"] = f"Bearer {bearer_token}"
-    r.headers["User-Agent"] = "v2FilteredStreamPython"
-    return r
-
-
-def get_rules():
-    response = requests.get(
-        "https://api.twitter.com/2/tweets/search/stream/rules", auth=bearer_oauth
-    )
-    if response.status_code != 200:
-        raise Exception(
-            "Cannot get rules (HTTP {}): {}".format(
-                response.status_code, response.text)
-        )
-    print(json.dumps(response.json()))
-    return response.json()
-
-
-def delete_all_rules(rules):
-    if rules is None or "data" not in rules:
-        return None
-
-    ids = list(map(lambda rule: rule["id"], rules["data"]))
-    payload = {"delete": {"ids": ids}}
-    response = requests.post(
-        "https://api.twitter.com/2/tweets/search/stream/rules",
-        auth=bearer_oauth,
-        json=payload
-    )
-    if response.status_code != 200:
-        raise Exception(
-            "Cannot delete rules (HTTP {}): {}".format(
-                response.status_code, response.text
-            )
-        )
-    print(json.dumps(response.json()))
-
-
-def set_rules(delete, update_flag):
-    # add more error handling for real-time rule adjustment gaps
-    with open("utils/yamls/rules.yml", "r") as file:
-        axel_rules = yaml.load(file, Loader=yaml.FullLoader)
-
-    print("RULES SAVED TO rules.yml")
-    print("UPDATE VALUE IN SET: ", update_flag)
-    if update_flag:
-        axel_rules = axel_rules + \
-            [{"value": config["ADD_RULE"], "tag": config["ADD_TAG"]}, ]
-        with open("utils/yamls/rules.yml", "w") as file:
-            file.write(str(axel_rules))
-
-        print("RULE VALUE UPDATED:\n", update_flag)
-        print(("ADDED RULES USED:\n", axel_rules))
-
-    # Reconnect stream if not active and set rules again
-    response = requests.get(
-        "https://api.twitter.com/2/tweets/search/stream/rules", auth=bearer_oauth
-    )
-    rules = get_rules()
-    if response.status_code != 200:
-        delete_all_rules(rules)
-        print("Reconnecting to the stream...")
-        with open("utils/yamls/config.yml", "w") as file:
-            config["RECONNECT_COUNT"] += 1
-            yaml.dump(config, file)
-
-    payload = {"add": axel_rules}
-    response = requests.post(
-        "https://api.twitter.com/2/tweets/search/stream/rules",
-        auth=bearer_oauth,
-        json=payload,
-    )
-    if response.status_code != 201:
-        raise Exception(
-            "Cannot add rules (HTTP {}): {}".format(
-                response.status_code, response.text)
-        )
-    print(json.dumps(response.json()))
-
-
-def update_rules():
-    with open("utils/yamls/config.yml", "r") as file:
-        config = yaml.load(file, Loader=yaml.FullLoader)
-
-    if "ADD_RULE" in config:
-        rule = config["ADD_RULE"]
-        update_flag = True
-        print("UPDATED TO TRUE: ", update_flag)
-    else:
-        print("No rule to add")
-
-    if rule == "":
-        update_flag = False
-        print("UPDATED TO FALSE: ", update_flag)
-    else:
-        print("SETTING NEW RULES")
-        delete = delete_all_rules(get_rules())
-
-        set_rules(delete, update_flag)
-        update_flag = False
-    with open("utils/yamls/config.yml", "w") as file:
-        config["ADD_RULE"] = ""
-        yaml.dump(config, file)
-        print("RULE RESET TO EMPTY")
-
-
-def remove_rules(rules):
-    remove_it = config["REMOVE_RULE"]
-    if remove_it == "":
-        print("NO RULE IN CONFIG TO REMOVE")
-        return None
-    new_rules = []
-    for rule in rules:
-        if rule["value"] != remove_it:
-            new_rules.append(rule)
-    print("NEW RULES: ", new_rules)
-    with open("utils/yamls/rules.yml", "w") as file:
-        file.write(str(new_rules))
-    with open("utils/yamls/config.yml", "w") as file:
-        config["REMOVE_RULE"] = ""
-        yaml.dump(config, file)
-        print("REMOVE RULE RESET TO EMPTY")
-
-    delete_all_rules(get_rules())
-    set_rules(new_rules, update_flag)
-    remove_flag = True
-    return remove_flag
-
-
-def get_data_by_id(tweet_id):
-    response = requests.get(
-        f"https://api.twitter.com/2/tweets/{tweet_id}?expansions=author_id,entities.mentions.username,geo.place_id,referenced_tweets.id&media.fields=url&poll.fields=options&tweet.fields=public_metrics",
-        auth=bearer_oauth
-    )
-    if response.status_code != 200:
-        raise Exception(
-            "Cannot get tweet data (HTTP {}): {}".format(
-                response.status_code, response.text)
-        )
-    return response.json()
-
-
-def get_likes_retweets_impressions(tweet_id):
-    response = requests.get(
-        f"https://api.twitter.com/1.1/statuses/show.json?id={tweet_id}",
-        auth=bearer_oauth
-    )
-
-    if response.status_code != 200:
-        raise Exception(
-            "Cannot get tweet data (HTTP {}): {}".format(
-                response.status_code, response.text)
-        )
-    return response.json()
-
-
-def get_username_by_author_id(author_id):
-    response = requests.get(
-        f"https://api.twitter.com/2/users/{author_id}",
-        auth=bearer_oauth
-    )
-    if response.status_code != 200:
-        raise Exception(
-            "Cannot get user data (HTTP {}): {}".format(
-                response.status_code, response.text)
-        )
-    return response.json()
+def get_export_df():
+    return export_df
 
 
 def get_stream(update_flag, remove_flag):
     response = requests.get(
-        "https://api.twitter.com/2/tweets/search/stream", auth=bearer_oauth, stream=True,
+        "https://api.twitter.com/2/tweets/search/stream", auth=st.bearer_oauth, stream=True,
     )
 
     print(response.status_code)
@@ -212,7 +44,7 @@ def get_stream(update_flag, remove_flag):
             with open("utils/yamls/config.yml", "w") as file:
                 config["RECONNECT_COUNT"] += 1
                 yaml.dump(config, file)
-            set_rules(delete_all_rules(get_rules()), update_flag)
+            st.set_rules(st.delete_all_rules(st.get_rules()), update_flag)
         except:
             raise Exception(
                 "Cannot get stream (HTTP {}): {}".format(
@@ -224,11 +56,11 @@ def get_stream(update_flag, remove_flag):
             print("\n\nGOT RESPONSE!")
             if update_flag:
                 print("UPDATING RULES")
-                update_rules()
+                st.update_rules()
                 update_flag = False
             if remove_flag:
                 print("REMOVING RULES")
-                remove_rules(get_rules())
+                st.remove_rules(st.get_rules())
                 remove_flag = False
 
             json_response = json.loads(response_line)
@@ -248,7 +80,7 @@ def get_stream(update_flag, remove_flag):
             # aggregate (x/x)*engagement to quote/retweeter
 
             print("\nTweet ID: ", id)
-            tweet_data = get_data_by_id(id)
+            tweet_data = st.get_data_by_id(id)
 
             # print raw data dump
             # print("\nDATA BY ID: ", json.dumps(
@@ -261,7 +93,7 @@ def get_stream(update_flag, remove_flag):
                 if tweet_data["data"]["author_id"]:
                     if "author_id" in tweet_data["data"]:
                         author_id = tweet_data["data"]["author_id"]
-                        author = get_username_by_author_id(
+                        author = st.get_username_by_author_id(
                             tweet_data["data"]["author_id"])
                         author_username = author["data"]["username"]
                         author_name = author["data"]["name"]
@@ -292,7 +124,7 @@ def get_stream(update_flag, remove_flag):
                 print("Restarting stream...")
                 get_stream(update_flag, remove_flag)
 
-            engagement_metrics = get_likes_retweets_impressions(id)
+            engagement_metrics = st.get_likes_retweets_impressions(id)
             tweet_favorite_count = int(engagement_metrics["favorite_count"])
             tweet_retweet_count = int(engagement_metrics["retweet_count"])
             print("\nTweet Favorites: ", tweet_favorite_count)
@@ -372,6 +204,7 @@ def get_stream(update_flag, remove_flag):
                                    data=id, columns=["Tweet ID"])
                 df = pd.concat([df0, df1, df2, df3], axis=1)
                 outputs_df = df
+                export_df = df
 
             # ,
             # outputs_df.to_csv("outputs/df.csv", sep="\t")
@@ -397,7 +230,7 @@ def get_stream(update_flag, remove_flag):
 
                     if included_author_id == author_id:
                         included_author_username = author_username
-                        included_user = get_username_by_author_id(
+                        included_user = st.get_username_by_author_id(
                             included_author_id)
                         included_author_name = included_user["data"]["name"]
                         outputs_df = pd.read_csv(
@@ -414,7 +247,7 @@ def get_stream(update_flag, remove_flag):
 
                     else:
                         try:
-                            inclued_name = get_username_by_author_id(
+                            inclued_name = st.get_username_by_author_id(
                                 included_author_id)
                             included_author_username = inclued_name["data"]["username"]
                             outputs_df = pd.read_csv(
@@ -446,16 +279,12 @@ def get_stream(update_flag, remove_flag):
                                     index=authors_index, data=included_id, columns=["Tweet ID"])
                                 df = pd.concat([df0, df1, df2, df3], axis=1)
                             outputs_df = outputs_df.append(df)
+                            export_df = df
                         except:
                             print("ERROR ON GET USERNAME BY AUTHOR ID")
-                        # outputs_df.to_csv("outputs/df.csv", sep="\t")s
 
                     print("\nAUTHOR OF INCLUDED/PARENT TWEET DIFFERENT FROM AUTHOR")
                     outputs_df.to_csv("outputs/df.csv", sep="\t")
-
-                    # comment becuase we are printing the members below
-                    # print("\nIncluded Tweet Public Metrics: ",
-                    #       included_pub_metrics)
 
                     print("\nIncluded/Parent Likes: ", included_likes)
                     print("\nIncluded/Parent Replies: ", included_reply_count)
@@ -480,6 +309,7 @@ def get_stream(update_flag, remove_flag):
                         # print("\nMatching Mentioned Author Name: ", engager_name)
                         # print(
                         #     "\nMatching Mentioned Author Username: ", engager_username)
+
                         engager_author_username = author_username
                         outputs_df = pd.read_csv(
                             "outputs/df.csv", index_col=0, on_bad_lines="skip")
@@ -547,10 +377,8 @@ def get_stream(update_flag, remove_flag):
                                 df = pd.concat([df1, df2, df3], axis=1)
 
                 outputs_df.to_csv("outputs/df.csv", sep="\t")
-                outputs_json.to_json("outputs/df.json", orient="index")
-
-                # print("Outputs_df Author: ",
-                #       outputs_df.loc["sora_joey", "Author"])
+                # outputs_json.to_json("outputs/df.json", orient="index")
+                export_df = df  # FIX THIS EXPORT DF SO USABLE IN POSTGRES_TOOLS.PY
 
                 # TODO: add logic to compare metrics for author and included/parent author
                 # aggregate stats for participating author + stats for included/parent author
@@ -558,9 +386,9 @@ def get_stream(update_flag, remove_flag):
 
 
 def main():
-    rules = get_rules()
-    delete = delete_all_rules(rules)
-    set = set_rules(delete, update_flag)
+    rules = st.get_rules()
+    delete = st.delete_all_rules(rules)
+    set = st.set_rules(delete, update_flag)
     get_stream(update_flag, remove_flag)
