+++ utils/postgres_tools.py
@@ -3,6 +3,7 @@ import psycopg2
 import csv
 import os
 import subprocess
+import filtered_stream as fs
 from sqlalchemy import create_engine
 from dotenv import load_dotenv
 load_dotenv()
@@ -100,12 +101,6 @@ POSTGRES_PASSWORD = os.getenv("POSTGRES_PASSWORD")
 #             cur.close()
 #             conn.close()
 
-
-# def start_db(db_name):
-#     engine = create_engine(
-#         'postgresql://'+POSTGRES_USER+':'+POSTGRES_PASSWORD+'@localhost:5432/'+db_name)
-#     return engine
-
 # def start_db(db_name):
 #     if check_postgres():
 #         engine = create_engine(
@@ -118,10 +113,6 @@ POSTGRES_PASSWORD = os.getenv("POSTGRES_PASSWORD")
 #         engine = create_engine(
 #             'postgresql://'+POSTGRES_USER+':'+POSTGRES_PASSWORD+'@localhost:5432/'+db_name)
 #     return engine
-
-
-# def write_to_db(engine, df, table_name):
-#     df.to_sql(table_name, engine, if_exists='append', index=True)
 
 # def write_to_db(engine, df, table_name):
 #     if check_postgres():
@@ -158,8 +149,10 @@ POSTGRES_PASSWORD = os.getenv("POSTGRES_PASSWORD")
 # if __name__ == "__main__":
 #     main()
 
+
 # SUBPROCESS VERSION
 
+# CSV FUNCTIONS
 def write_csv(df, csv_path):
     df.to_csv(csv_path, index=False, quoting=csv.QUOTE_ALL)
 
@@ -169,26 +162,17 @@ def read_csv(csv_path):
         csv_path, index_col=0, on_bad_lines="skip")
     return df
 
-
-def get_admin_user(database_name):
-    conn = psycopg2.connect(
-        host="localhost",
-        database=database_name,
-        user="postgres",
-        password="postgres"
-    )
-    cursor = conn.cursor()
-    cursor.execute("SELECT * FROM pg_user")
-    rows = cursor.fetchall()
-    for row in rows:
-        print(row)
+# DATABASE ADMIN FUNCTIONS
 
 
-admin_user = get_admin_user("test")
-if admin_user:
-    print(f"The administrator user is: {admin_user}")
-else:
-    print("Failed to get the administrator user.")
+def create_db(db_name, user):
+    try:
+        print("\nCreating database...")
+        subprocess.run(["createdb", "-U", POSTGRES_ADMIN_USER,
+                        "-O", user, db_name], check=True)
+        print("Database created")
+    except subprocess.CalledProcessError:
+        print("Database '{}' already exists".format(db_name))
 
 
 def create_role(user, password):
@@ -202,20 +186,47 @@ def create_role(user, password):
     except subprocess.CalledProcessError:
         print("Role '{}' already exists.".format(user))
 
+# DATABASE INTERACTION FUNCTIONS
 
-def create_db(db_name, user):
-    try:
-        print("\nCreating database...")
-        subprocess.run(["createdb", "-U", POSTGRES_ADMIN_USER,
-                        "-O", user, db_name], check=True)
-        print("Database created")
-    except subprocess.CalledProcessError:
-        print("Database '{}' already exists".format(db_name))
+
+def start_db(db_name):
+    engine = create_engine(
+        'postgresql://'+POSTGRES_USER+':'+POSTGRES_PASSWORD+'@localhost:5433/'+db_name)
+    return engine
+
+
+def write_to_db(engine, df, table_name):
+    df.to_sql(table_name, engine, if_exists='append', index=True)
+
+
+def get_admin_user(database_name):
+    conn = psycopg2.connect(
+        host="localhost",
+        database=database_name,
+        user="postgres",
+        password=POSTGRES_PASSWORD
+    )
+    cursor = conn.cursor()
+    cursor.execute("SELECT * FROM pg_user")
+    rows = cursor.fetchall()
+    for row in rows:
+        print(row)
 
 
 def main():
-    create_role(POSTGRES_USER, POSTGRES_PASSWORD)
-    create_db("test", POSTGRES_ADMIN_USER)
+    # Assuming database and roles aren't created yet
+    # create_role(POSTGRES_USER, POSTGRES_PASSWORD)
+    # create_db("test", POSTGRES_ADMIN_USER)
+
+    # TODO: fix logic for getting export_df from fs to write to database
+    # we will either call this in a listener that triggers with each new twitter listener trigger
+    # or we will call this in a while loop that refreshes every x minutes and pushes DF updates to database
+
+    # Assuming database and roles are created
+    engine = start_db("test")
+    df = fs.get_export_df()
+    print(df)
+    write_to_db(engine, df, "df_table")
 
 
 if __name__ == "__main__":
