+++ utils/stream_tools.py
@@ -1,0 +1,199 @@
+import json
+import os
+import requests
+import yaml
+import filtered_stream as fs
+from dotenv import load_dotenv
+load_dotenv()
+
+with open("utils/yamls/config.yml", "r") as file:
+    config = yaml.load(file, Loader=yaml.FullLoader)
+
+bearer_token = os.environ.get("TWITTER_BEARER_TOKEN")
+
+
+def bearer_oauth(r):
+    """
+    Method required by bearer token authentication.
+    """
+
+    r.headers["Authorization"] = f"Bearer {bearer_token}"
+    r.headers["User-Agent"] = "v2FilteredStreamPython"
+    return r
+
+
+def get_username_by_author_id(author_id):
+    response = requests.get(
+        f"https://api.twitter.com/2/users/{author_id}",
+        auth=bearer_oauth
+    )
+    if response.status_code != 200:
+        raise Exception(
+            "Cannot get user data (HTTP {}): {}".format(
+                response.status_code, response.text)
+        )
+    return response.json()
+
+
+def get_rules():
+    response = requests.get(
+        "https://api.twitter.com/2/tweets/search/stream/rules", auth=bearer_oauth
+    )
+    if response.status_code != 200:
+        raise Exception(
+            "Cannot get rules (HTTP {}): {}".format(
+                response.status_code, response.text)
+        )
+    print(json.dumps(response.json()))
+    return response.json()
+
+
+def delete_all_rules(rules):
+    if rules is None or "data" not in rules:
+        return None
+
+    ids = list(map(lambda rule: rule["id"], rules["data"]))
+    payload = {"delete": {"ids": ids}}
+    response = requests.post(
+        "https://api.twitter.com/2/tweets/search/stream/rules",
+        auth=bearer_oauth,
+        json=payload
+    )
+    if response.status_code != 200:
+        raise Exception(
+            "Cannot delete rules (HTTP {}): {}".format(
+                response.status_code, response.text
+            )
+        )
+    print(json.dumps(response.json()))
+
+
+def set_rules(delete, update_flag):
+    # add more error handling for real-time rule adjustment gaps
+    with open("utils/yamls/rules.yml", "r") as file:
+        axel_rules = yaml.load(file, Loader=yaml.FullLoader)
+
+    print("RULES SAVED TO rules.yml")
+    print("UPDATE VALUE IN SET: ", update_flag)
+    if update_flag:
+        axel_rules = axel_rules + \
+            [{"value": config["ADD_RULE"], "tag": config["ADD_TAG"]}, ]
+        with open("utils/yamls/rules.yml", "w") as file:
+            file.write(str(axel_rules))
+
+        print("RULE VALUE UPDATED:\n", update_flag)
+        print(("ADDED RULES USED:\n", axel_rules))
+
+    # Reconnect stream if not active and set rules again
+    response = requests.get(
+        "https://api.twitter.com/2/tweets/search/stream/rules", auth=bearer_oauth
+    )
+    rules = get_rules()
+    if response.status_code != 200:
+        delete_all_rules(rules)
+        print("Reconnecting to the stream...")
+        with open("utils/yamls/config.yml", "w") as file:
+            config["RECONNECT_COUNT"] += 1
+            yaml.dump(config, file)
+
+    payload = {"add": axel_rules}
+    response = requests.post(
+        "https://api.twitter.com/2/tweets/search/stream/rules",
+        auth=bearer_oauth,
+        json=payload,
+    )
+    if response.status_code != 201:
+        raise Exception(
+            "Cannot add rules (HTTP {}): {}".format(
+                response.status_code, response.text)
+        )
+    print(json.dumps(response.json()))
+
+
+def update_rules():
+    with open("utils/yamls/config.yml", "r") as file:
+        config = yaml.load(file, Loader=yaml.FullLoader)
+
+    if "ADD_RULE" in config:
+        rule = config["ADD_RULE"]
+        update_flag = True
+        print("UPDATED TO TRUE: ", update_flag)
+    else:
+        print("No rule to add")
+
+    if rule == "":
+        update_flag = False
+        print("UPDATED TO FALSE: ", update_flag)
+    else:
+        print("SETTING NEW RULES")
+        delete = delete_all_rules(get_rules())
+
+        set_rules(delete, update_flag)
+        update_flag = False
+    with open("utils/yamls/config.yml", "w") as file:
+        config["ADD_RULE"] = ""
+        yaml.dump(config, file)
+        print("RULE RESET TO EMPTY")
+
+
+def remove_rules(rules):
+    remove_it = config["REMOVE_RULE"]
+    if remove_it == "":
+        print("NO RULE IN CONFIG TO REMOVE")
+        return None
+    new_rules = []
+    for rule in rules:
+        if rule["value"] != remove_it:
+            new_rules.append(rule)
+    print("NEW RULES: ", new_rules)
+    with open("utils/yamls/rules.yml", "w") as file:
+        file.write(str(new_rules))
+    with open("utils/yamls/config.yml", "w") as file:
+        config["REMOVE_RULE"] = ""
+        yaml.dump(config, file)
+        print("REMOVE RULE RESET TO EMPTY")
+
+    delete_all_rules(get_rules())
+    set_rules(new_rules, fs.update_flag)
+    remove_flag = True
+    return remove_flag
+
+
+def get_data_by_id(tweet_id):
+    response = requests.get(
+        f"https://api.twitter.com/2/tweets/{tweet_id}?expansions=author_id,entities.mentions.username,geo.place_id,referenced_tweets.id&media.fields=url&poll.fields=options&tweet.fields=public_metrics",
+        auth=bearer_oauth
+    )
+    if response.status_code != 200:
+        raise Exception(
+            "Cannot get tweet data (HTTP {}): {}".format(
+                response.status_code, response.text)
+        )
+    return response.json()
+
+
+def get_username_by_author_id(author_id):
+    response = requests.get(
+        f"https://api.twitter.com/2/users/{author_id}",
+        auth=bearer_oauth
+    )
+    if response.status_code != 200:
+        raise Exception(
+            "Cannot get user data (HTTP {}): {}".format(
+                response.status_code, response.text)
+        )
+    return response.json()
+
+
+def get_likes_retweets_impressions(tweet_id):
+    response = requests.get(
+        f"https://api.twitter.com/1.1/statuses/show.json?id={tweet_id}",
+        auth=bearer_oauth
+    )
+
+    if response.status_code != 200:
+        raise Exception(
+            "Cannot get tweet data (HTTP {}): {}".format(
+                response.status_code, response.text)
+        )
+    return response.json()
